third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/settings/geolocations-a11y-test-expected.txt

Tests accessibility in the settings tool geolocations pane using the axe-core linter.

Running: testAddLocation
Opened input box: true
aXe violations: []


Running: testNewLocationError
Invalidating the Location name input
Error message: Location name cannot be empty
Invalidating the Latitude input
Error message: Latitude must be a number
Invalidating the Longitude input
Error message: Longitude must be a number
aXe violations: []
